Complete Set 1 task 1 hex to base64
This took me a surprisingly long time. I think this is the first time
I've sat down and looked at the actual algorithm for converting either
from hex to bytes, or from bytes to base64. Makes sense now I've wrapped
my head around it, and I've tried to leave a good comment on my
understanding of the bit shifting that makes this work.

Fun to do some stuff with bits and bytes for a change!