Set 1 challenge 3 - single byte XOR cipher
This was fun. Had to figure out the best way to score a piece of english
text. At first I started by just counting `e`s, assuming that's the most
common char. Then realised that it would be much better to improve a
score for having A-Za-z and punishing for chars not in that range. Was
off in my ASCII lookup but correcting fixed that.