Set 1 Challenge 5
Spent a little while trying to figure out what weird stuff was going on
with newlines in this challenge. Ended up being that their input text
had a new line char (which was to be xored normally) but the result text
was meant to have no newline. That was a bit confusing, but got there in
the end.

Should do the repeating xor more to get a feel for it.